backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
anthropic==0.73.0
PyMuPDF==1.26.6
openai==1.55.3
Pillow==11.0.0
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
